refactor(lint): Phase 2 mechanical clippy fixes
cargo clippy --fix로 일괄 적용한 의미 변경 없는 스타일 정정 + 수동 보정 7건.

[--fix 자동 적용 — 약 150건]
- collapsible_if (~28): 중첩 if를 let-chain (`&&` 연결)으로 축소
- explicit_auto_deref (~31): 잉여 `&*x` deref 제거
- single_char_add_str (~17): `s.push_str("\n")` → `s.push('\n')`
- unnecessary_cast (~6): `x as f32` (x already f32) 제거
- unnecessary_map_or (~3): `map_or(true, P)` → `is_none_or(P)`
- expect_fun_call (~4): `expect(&format!(...))` → `unwrap_or_else(|| panic!(...))`
- print_literal (~6), useless_format (~2), redundant_closure (~2),
  bool_assert_comparison (~2), len_zero (~1), needless_borrow (~1),
  manual_range_contains (~2)

[수동 보정 — --fix가 처리 못한 7건]
- handler_v2.rs · mcp_server.rs: field_reassign_with_default →
  struct literal + `..Default::default()` 형태로 (테스트/시나리오 로드).
- llama_timings_test.rs: `while let Some(_) = stream.next().await {}` →
  `while stream.next().await.is_some() {}` (redundant_pattern_matching).
- magnitude_bench.rs: 모듈 docstring 리스트 들여쓰기 수정 (overindent).
- listener_perspective_integration_bench.rs: `&"...".to_string()` →
  `"..."` (push_str은 &str 직접 수용).
- dialogue_orchestrator_test.rs · memory_injection_test.rs:
  복합 튜플 반환 시그니처를 type alias (SetupTuple/SetupWithMemoryTuple)로.

[제외]
assertions_on_constants 43건은 의미 격상(런타임 → 컴파일타임)이라
별도 커밋으로 분리.

검증: cargo test (default) 270+건 통과. full-feature clippy
잔여 경고 = assertions_on_constants 43건만.

+++ tests/personality_test.rs
@@ -655,8 +655,8 @@ fn test_derive_llm_parameters() {
     
     // Formula: temp = 0.8 + (1.0*0.1) + (1.0*0.05) - (-1.0*0.1) - (-1.0*0.05) = 0.8 + 0.1 + 0.05 + 0.1 + 0.05 = 1.1
     // Formula: top_p = 0.9 + (1.0*0.05) - (-1.0*0.05) = 0.9 + 0.05 + 0.05 = 1.0 (clamped to 1.0)
-    assert!((temp1 as f32 - 1.1).abs() < 0.001, "Expected temp 1.1, got {}", temp1);
-    assert!((top_p1 as f32 - 1.0).abs() < 0.001, "Expected top_p 1.0, got {}", top_p1);
+    assert!((temp1 - 1.1).abs() < 0.001, "Expected temp 1.1, got {}", temp1);
+    assert!((top_p1 - 1.0).abs() < 0.001, "Expected top_p 1.0, got {}", top_p1);
 
     // 2. Rigid profile: Low O, Low X, High C -> Expect low temperature
     let mut rigid_profile = HexacoProfile::neutral();
@@ -685,12 +685,12 @@ fn test_derive_llm_parameters() {
     
     // Formula: temp = 0.8 + (-1.0*0.1) + (-1.0*0.05) - (1.0*0.1) - (1.0*0.05) = 0.8 - 0.1 - 0.05 - 0.1 - 0.05 = 0.5
     // Formula: top_p = 0.9 + (-1.0*0.05) - (1.0*0.05) = 0.9 - 0.05 - 0.05 = 0.8
-    assert!((temp2 as f32 - 0.5).abs() < 0.001, "Expected temp 0.5, got {}", temp2);
-    assert!((top_p2 as f32 - 0.8).abs() < 0.001, "Expected top_p 0.8, got {}", top_p2);
+    assert!((temp2 - 0.5).abs() < 0.001, "Expected temp 0.5, got {}", temp2);
+    assert!((top_p2 - 0.8).abs() < 0.001, "Expected top_p 0.8, got {}", top_p2);
 
     // 3. Neutral profile
     let neutral_profile = HexacoProfile::neutral();
     let (temp3, top_p3) = neutral_profile.derive_llm_parameters();
-    assert!((temp3 as f32 - 0.8).abs() < 0.001, "Expected temp 0.8, got {}", temp3);
-    assert!((top_p3 as f32 - 0.9).abs() < 0.001, "Expected top_p 0.9, got {}", top_p3);
+    assert!((temp3 - 0.8).abs() < 0.001, "Expected temp 0.8, got {}", temp3);
+    assert!((top_p3 - 0.9).abs() < 0.001, "Expected top_p 0.9, got {}", top_p3);
 }
